requirements.txt
Flask
flask-cors
pytesseract
Pillow
spacy
numpy
nltk
